use hqcase/by_owner instead of case/by_owner in dbaccessors
this forces callers to specify a domain; this just seems safer,
and means we can probably soon get rid of case/by_owner
