+++ corehq/apps/hqadmin/views.py
@@ -32,11 +32,11 @@ from django.http import (
 )
 from restkit import Resource
 from restkit.errors import Unauthorized
-from casexml.apps.case.dbaccessors import get_total_case_count
 
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.callcenter.indicator_sets import CallCenterIndicators
 from couchdbkit import ResourceNotFound, Database
+from corehq.apps.hqcase.dbaccessors import get_total_case_count
 from corehq.apps.hqcase.utils import get_case_by_domain_hq_user_id
 from couchforms.const import DEVICE_LOG_XMLNS
 from couchforms.dbaccessors import get_number_of_forms_all_domains_in_couch
