+++ corehq/ex-submodules/casexml/apps/case/dbaccessors/__init__.py
@@ -1,5 +1,4 @@
 from .all_cases import *
-from .by_owner import *
 from .related import *
