+++ corehq/ex-submodules/casexml/apps/case/tests/test_bugs.py
@@ -2,7 +2,6 @@ import uuid
 from django.test import TestCase
 import os
 from django.test.utils import override_settings
-from casexml.apps.case.dbaccessors import get_total_case_count
 from casexml.apps.case.exceptions import IllegalCaseId
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.models import CommCareCase
@@ -11,6 +10,7 @@ from casexml.apps.case.templatetags.case_tags import get_case_hierarchy
 from casexml.apps.case.tests.util import delete_all_cases
 from casexml.apps.case.util import post_case_blocks
 from casexml.apps.case.xml import V2
+from corehq.apps.hqcase.dbaccessors import get_total_case_count
 from couchforms.tests.testutils import post_xform_to_couch
 from casexml.apps.case.xform import process_cases
