+++ corehq/ex-submodules/casexml/apps/case/tests/test_db_accessors.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
-from casexml.apps.case.dbaccessors import get_total_case_count
 from casexml.apps.case.models import CommCareCase
 from casexml.apps.case.tests import delete_all_cases
+from corehq.apps.hqcase.dbaccessors import get_total_case_count
 
 
 class TestCaseByOwner(TestCase):
