+++ corehq/ex-submodules/casexml/apps/phone/data_providers/case/clean_owners.py
@@ -2,13 +2,13 @@ from collections import defaultdict
 from copy import copy
 from functools import partial
 from datetime import datetime
-from casexml.apps.case.dbaccessors import get_open_case_ids
 from casexml.apps.case.models import CommCareCase
 from casexml.apps.phone.cleanliness import get_case_footprint_info
 from casexml.apps.phone.data_providers.case.load_testing import append_update_to_response
 from casexml.apps.phone.data_providers.case.stock import get_stock_payload
 from casexml.apps.phone.data_providers.case.utils import get_case_sync_updates, CaseStub
 from casexml.apps.phone.models import OwnershipCleanlinessFlag, LOG_FORMAT_SIMPLIFIED, IndexTree, SimplifiedSyncLog
+from corehq.apps.hqcase.dbaccessors import get_open_case_ids
 from corehq.apps.users.cases import get_owner_id
 from corehq.dbaccessors.couchapps.cases_by_server_date.by_owner_server_modified_on import \
     get_case_ids_modified_with_owner_since
@@ -127,7 +127,7 @@ class CleanOwnerCaseSyncOperation(object):
         if self.is_clean(owner_id):
             if self.restore_state.is_initial:
                 # for a clean owner's initial sync the base set is just the open ids
-                return set(get_open_case_ids(owner_id))
+                return set(get_open_case_ids(self.restore_state.domain, owner_id))
             else:
                 # for a clean owner's steady state sync, the base set is anything modified since last sync
                 return set(get_case_ids_modified_with_owner_since(
